grayscope/backend/requirements.txt
# GrayScope Backend Dependencies (Phase 1)
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.8.0
pydantic-settings>=2.4.0
sqlalchemy>=2.0.30
alembic>=1.13.0
httpx>=0.27.0
pyyaml>=6.0.1
jinja2>=3.1.4
tree-sitter>=0.23.0,<0.24.0
tree-sitter-c>=0.21.0,<0.22.0
tree-sitter-cpp>=0.22.0,<0.23.0
python-dotenv>=1.0.1
docker>=7.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
